feat: update components and home-page
